### 2017-05-18 #262 @whedon
[PRE REVIEW]: root_numpy: The interface between ROOT and NumPy
**Submitting author:** @ndawe (<a href="http://orcid.org/0000-0003-0202-3284">Edmund Dawe</a>)
**Repository:** <a href="https://github.com/scikit-hep/root_numpy" target ="_blank">https://github.com/scikit-hep/root_numpy</a>
**Version:** 4.7.2
