### 2017-05-19 c4f4d5d
DOCS: selection clarification for selection with multiplicity
