### 2017-05-19 78b19e1
DOCS: clarify object_selection
