### 2017-07-04 #262 @arfon
@openjournals/joss-editors - would anyone else be able to handle this submission for JOSS?
