### 2017-07-06 #262 @danielskatz
I'm looking for reviewer volunteers now.  @betatim, @jpivarski, and @wlav have all been suggested.  Would one of you be willing to review this paper for JOSS?  (for more info about the journal and process, see http://joss.theoj.org/about, and particularly http://joss.theoj.org/about#reviewer_guidelines)
