### 2017-07-06 #262 @wlav
Sure can do, assuming that I pass the COI threshold.
