### 2017-07-06 #262 @danielskatz
@wlav
Did you contribute to the code?
What other potential COIs should we be concerned with?
