### 2017-07-06 #262 @wlav
No, I did not contribute nor did I use the code. However, root_numpy uses PyROOT extensively, and I was the main developer of that. Thus, I very much like root_numpy to succeed.
