### 2017-07-10 scikit-hep/root_numpy#328 @wlav
joss review: error in tests, but test "ok"
When running the tests, I get:

root_numpy.tests.test_evaluate.test_evaluate_func` ... input_line_99:1:55: error: use of undeclared identifier 'f'
Double_t TFormula____id6180733986148604910(){ return {f} ; }
                                                      ^

And yet the result is:

ok

The 'error' may well be ignorable, but recommend to then suppress it. That said, the code really isn't valid C++, so something may be lurking underneath and thus the test should not have succeeded.
