### 2017-07-10 scikit-hep/root_numpy#332 @wlav
joss review: API reference discrepancies
``RootNumpyUnconvertibleWarning`` is part of ``__all__`` but not listed in the API reference

the function setup_utils() is public in dir() but not part of ``__all__`` or documented in the API reference (I don't think it should be?)
