### 2017-07-11 scikit-hep/root_numpy#328 @ndawe
Thanks @wlav for starting the JOSS review! I'll try to respond to each issue as quickly as possible, but I'll be back in Melbourne on July 27 so I might not be able to respond to all of them until after that.

This particular error is expected and is the result of the test where the function cannot be evaluated:

```python
assert_raises(ValueError, rnp.evaluate, "f", arr_1d)
```

For example:

```python
>>> import root_numpy as rnp
>>> import numpy as np
>>> rnp.evaluate("f", np.ones(10))
input_line_35:1:55: error: use of undeclared identifier 'f'
Double_t TFormula____id6180733986148604910(){ return {f} ; }
                                                      ^
Error in <TFormula::Eval>: Can't find TFormula____id6180733986148604910 function prototype with arguments
Error in <TFormula::ProcessFormula>: "f" has not been matched in the formula expression
Error in <TFormula::ProcessFormula>: Formula "f" is invalid !
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/endw/.local/lib/python2.7/site-packages/root_numpy/_evaluate.py", line 108, in evaluate
    raise ValueError("invalid formula expression")
ValueError: invalid formula expression
```

This seems to be an error originating within Cling. Do you know of a way to silence Cling from writing to stdout/stderr?
