### 2017-07-12 scikit-hep/root_numpy#328 @wlav
I'm not aware of a public API to do that and passing the Cling redirection syntax on gInterpreter.ProcessLine() seems to have no effect. However, you can temporarily redirect stderr output to /dev/null,  Example:

```python
import os, sys
import root_numpy as rnp
import numpy as np
save = file( '/dev/null', 'w' )
os.dup2(sys.stderr.fileno(), save.fileno())
shutup = file( '/dev/null', 'w' )
print >> sys.stderr, 'loud'
os.dup2(shutup.fileno(), sys.stderr.fileno())
print >> sys.stderr, 'silent'
try:
   rnp.evaluate("f", np.ones(10))
except ValueError:
   pass
os.dup2(save.fileno(), sys.stderr.fileno())
print >> sys.stderr, 'loud'
```

Of course, this silences all errors, not just the ones you expect or only the ones from Cling, so capturing the output in an unlinked file for printing if need be, may serve better.
