### 2017-07-17 #307 @danielskatz
@wlav - I see that you've checked off a bunch of thing, but not all of them.  Are the ones that are not checked items you are confirming?  Or have you been opening issues in https://github.com/scikit-hep/root_numpy/issues to discuss them (as it appears)?  If the latter, can you also put pointers to the review issues here so that others can look back and see the review progress?
