### 2017-07-17 #307 @wlav
Yes; there are several issues opened (mostly minor stuff except one or two). But @ndawe isn't back until July 27th, so waiting for that event. Will be soon enough.

This one is important: one item missing in the API reference that is exported with `__all__`
[https://github.com/scikit-hep/root_numpy/issues/332](url)

And this would be nice to have fixed/improved (error print, even if that is the 'ok' result for the test), but may be clunky to do:
[https://github.com/scikit-hep/root_numpy/issues/328](url)

There are three more issues that are no more than clarifications, and two where I misunderstood what was going on (e.g. build warnings coming from cython, not root_numpy).
