### 2017-07-18 b81d80e
FIX #328: silence cling error in test_evaluate.py
