### 2017-07-18 scikit-hep/root_numpy#328 @ndawe
Thanks @wlav! I've fixed this in #335
