### 2017-07-18 719c5f4
FIX #330: use testdata.get_filepath()
