+++ docs/start.rst
@@ -56,9 +56,9 @@ following examples into your Python prompt):
 .. code-block:: python
 
    from root_numpy import root2array, tree2array
-   from root_numpy.testdata import get_filepath
+   from root_numpy import testdata
 
-   filename = get_filepath('test.root')
+   filename = testdata.get_filepath('test.root')
 
    # Convert a TTree in a ROOT file into a NumPy structured array
    arr = root2array(filename, 'tree')
