### 2017-08-07 f547e47
Document RootNumpyUnconvertibleWarning. Fix #332
