### 2017-08-07 scikit-hep/root_numpy#332 @ndawe
Fixed in #339. Right, setup_utils is for internal use only.
