### 2017-08-07 f11d8c5
[ci skip] clarify running nosetests outside source directory and make test inside source directory
