### 2017-08-07 a025754
DOCS: promote Contributing section to top-level and make stackoverflow the recommended way to ask general questions
