+++ docs/start.rst
@@ -131,23 +131,3 @@ or histogram:
    hist.FillRandom('gaus')
    arr = random_sample(hist, 1000000)
 
-
-Have Questions or Found a Bug?
-==============================
-
-Think you found a bug? Open a new issue here:
-`github.com/rootpy/root_numpy/issues <https://github.com/rootpy/root_numpy/issues>`_.
-
-Also feel free to post questions or follow discussion on the
-`rootpy-users <http://groups.google.com/group/rootpy-users>`_ or
-`rootpy-dev <http://groups.google.com/group/rootpy-dev>`_ Google groups.
-
-
-Contributing
-============
-
-Please post on the rootpy-dev@googlegroups.com list if you have ideas
-or contributions. Feel free to fork
-`root_numpy on GitHub <https://github.com/rootpy/root_numpy>`_
-and later submit a pull request.
-
