### 2017-08-07 f906d45
FAQ: remove use of rootpy
