### 2017-08-07 #307 @danielskatz
@ndawe - any update from you on addressing the issues raised in the initial review?
