### 2017-08-07 #307 @ndawe
@danielskatz I've finally addressed the remaining issues:

https://github.com/scikit-hep/root_numpy/issues?utf8=%E2%9C%93&q=is%3Aissue%20label%3AJOSS%20

 @wlav is certainly welcome to open more if needed
