### 2017-08-11 146c9d9
JOSS paper: fix bib references (DOI) and minor text edits
