+++ paper/paper.bib
@@ -1,24 +1,53 @@
-@online{NumPy,
-  author = {NumPy},
-  title = {www.numpy.org},
-  year = 2017,
-  url = {http://www.numpy.org},
-  urldate = {2017-05-11}
+@article{NumPy,
+  author    = {Stéfan van der Walt and S. Chris Colbert and Gaël Varoquaux},
+  title     = {The NumPy Array: A Structure for Efficient Numerical Computation},
+  journal   = {Computing in Science \& Engineering},
+  volume    = {13},
+  pages     = {22-30},
+  year      = {2011},
+  doi       = {10.1109/MCSE.2011.37},
+  url       = {http://www.numpy.org},
 }
 
-@online{ROOT,
-  author = {ROOT},
-  title = {root.cern.ch},
-  year = 2017,
-  url = {https://root.cern.ch},
-  urldate = {2017-05-11}
+@article{ROOT,
+  author    = {Ilka Antcheva and Maarten Ballintijn and Bertrand Bellenot and Marek Biskup and Rene Brun and Nenad Buncic and Philippe Canal and Diego Casadei and Olivier Couet and Valery Fine and Leandro Franco and Gerardo Ganis and Andrei Gheata and David Gonzalez Maline and Masaharu Goto and Jan Iwaszkiewicz and Anna Kreshuk and Diego Marcos Segura and Richard Maunder and Lorenzo Moneta and Axel Naumann and Eddy Offermann and Valeriy Onuchin and Suzanne Panacek and Fons Rademakers and Paul Russo and Matevz Tadel},
+  title     = {ROOT - A C++ Framework for Petabyte Data Storage, Statistical Analysis and Visualization},
+  journal   = {Computer Physics Communications},
+  volume    = {180},
+  number    = {12},
+  pages     = {2499-2512},
+  year      = {2009},
+  doi       = {10.1016/j.cpc.2009.08.005},
+  url       = {https://root.cern.ch},
+}
+
+@article{Cython,
+  author    = {Stefan Behnel and Robert Bradshaw and Craig Citro and Lisandro Dalcin and Dag Sverre Seljebotn and Kurt Smith},
+  title     = {Cython: The Best of Both Worlds},
+  journal   = {Computing in Science and Engineering},
+  volume    = {13},
+  pages     = {31-39},
+  year      = {2011},
+  doi       = {10.1109/MCSE.2010.118},
+  url       = {http://cython.org}
 }
 
 @article{TMVA,
-	author = "Hoecker, Andreas and Speckmayer, Peter and Stelzer, Joerg and Therhaag, Jan and von Toerne, Eckhard and Voss, Helge",
-	title = "{TMVA: Toolkit for Multivariate Data Analysis}",
-	journal = "PoS",
-	volume = "ACAT",
-	year = "2007",
-	pages = "040",
+	author    = {Speckmayer, P. and Hocker, A. and Stelzer, J. and Voss, H.},
+	title     = {The toolkit for multivariate data analysis, TMVA 4},
+	journal   = {J. Phys. Conf. Ser.},
+	volume    = {219},
+	year      = {2010},
+	pages     = {032057},
+	doi       = {10.1088/1742-6596/219/3/032057},
+}
+
+@article{scikit-learn,
+  author    = {Fabian Pedregosa and Gaël Varoquaux and Alexandre Gramfort and Vincent Michel and Bertrand Thirion and Olivier Grisel and Mathieu Blondel and Peter Prettenhofer and Ron Weiss and Vincent Dubourg and Jake Vanderplas and Alexandre Passos and David Cournapeau and Matthieu Brucher and Matthieu Perrot and Édouard Duchesnay},
[…]
