### 2017-08-11 #307 @wlav
@ndawe: I'm posting this whole thing even as it may be a bit pointless, but enough digging has gone into it. On the plus side, I now know everything there is to know about root_numpy ...

Anyway ...

I had a hard time trying to reproduce the performance plots in the paper. Initially, I got this:

![bench_tree2array_v6 06 02](https://user-images.githubusercontent.com/6372198/29198225-d0763f94-7df6-11e7-855d-9ab44bcf3d58.png)

Digging around, the two codes do essentially the same thing, but the setup of ROOT is heavier (because of generality), so that would explain the first part of the left plot as I found it. But for the second part, I don't see why root_numpy would be lagging. The cython generated code seems efficient enough as concerns the inner loop: there is no cython or python specific overhead in there.

For the right plot, it is all self consistent: if the overhead per entry (i.e. in the inner loop) is greater, the linear scaling factor will be higher. And indeed in both your and my case, the right follows the left.

So .... somewhere there was some per-call overhead left. Turns out, that the default root_numpy does not have the optimization flags on. I added -O3, et voila:

![bench_tree2array_v6 06 02_2](https://user-images.githubusercontent.com/6372198/29198975-7e43cffa-7dfd-11e7-8df9-15fdd7ef6aba.png)

Mind you, I have a slower disk (it's a 5 year old spinning one), so convergence as seen on the left is expected, and the relative overhead as seen on the right also expected to be less (but diverging in the same direction nonetheless).

You may want to guarantee that compiler optimizations are on when building root_numpy, or warn that they have not been when doing the benchmark. I'll file an issue on that one; do with it as you like.

Aside, instead of sysinfo (which didn't work for me as the ioctl call failed), I just read /proc/scsi/scsi, which does not require uid==0.
