### 2017-08-11 a7cb9f0
mention scikit-learn and tensorflow
