+++ paper/paper.bib
@@ -51,3 +51,49 @@
   year      = {2011},
   url       = {http://scikit-learn.org},
 }
+
+@misc{TensorFlow,
+  title     = {{TensorFlow}: Large-Scale Machine Learning on Heterogeneous Systems},
+  url       = {http://tensorflow.org/},
+  author    = {Martín~Abadi and
+               Ashish~Agarwal and
+               Paul~Barham and
+               Eugene~Brevdo and
+               Zhifeng~Chen and
+               Craig~Citro and
+               Greg~S.~Corrado and
+               Andy~Davis and
+               Jeffrey~Dean and
+               Matthieu~Devin and
+               Sanjay~Ghemawat and
+               Ian~Goodfellow and
+               Andrew~Harp and
+               Geoffrey~Irving and
+               Michael~Isard and
+               Yangqing Jia and
+               Rafal~Jozefowicz and
+               Lukasz~Kaiser and
+               Manjunath~Kudlur and
+               Josh~Levenberg and
+               Dan~Mané and
+               Rajat~Monga and
+               Sherry~Moore and
+               Derek~Murray and
+               Chris~Olah and
+               Mike~Schuster and
+               Jonathon~Shlens and
+               Benoit~Steiner and
+               Ilya~Sutskever and
+               Kunal~Talwar and
+               Paul~Tucker and
+               Vincent~Vanhoucke and
+               Vijay~Vasudevan and
+               Fernanda~Viégas and
+               Oriol~Vinyals and
+               Pete~Warden and
+               Martin~Wattenberg and
+               Martin~Wicke and
+               Yuan~Yu and
+               Xiaoqiang~Zheng},
+  year      = {2015},
+}
