### 2017-08-11 #307 @danielskatz
@wlav - thanks for this very interesting analysis.  We may need to create a new award for most thorough review :)

Are you asking for changes in the repo/paper regarding this, or is this more optional for the author: observation and feedback to improve the work, but not required for JOSS acceptance?

And, you did not check the References box in the checklist.  Is this because the scikit-learn and Tensor Flow references do not have DOIs?  Or something else?
