### 2017-08-11 #307 @ndawe
@wlav sees a closer match between ROOT and root_numpy leading me to suspect I have been comparing root_numpy with compiler optimisations enabled to ROOT without optimisation. I'll confirm and recreate the plot if needed.
