### 2017-08-11 #307 @danielskatz
thanks @ndawe  - can you also check to see if there are DOIs that should be added for the last 2 entries in the bib file associated with the paper, so we can check off that item as well?
