### 2017-08-11 #307 @wlav
@danielkatz: There's an arXiv for TensorFlow: abs/1603.04467 (https://arxiv.org/pdf/1610.06918.pdf; and no, I don't know why the numbers differ). I can't find a doi. The citation for scikit-learn is the way they want it (see: http://scikit-learn.org/stable/about.html).
