### 2017-08-11 #307 @danielskatz
Ok, given no DOI for either, I think you can check off the references item, since "all archival references that should have a DOI" do list one; the ones that don't can't since there isn't one.
