### 2017-08-11 #307 @danielskatz
@ndawe, please let me know when you are satisfied with the state of the documentation and plot (as discussed by you and @wlav above).  If you need to make changes, please do.

Then we can go ahead and accept this, once you say you are ready.
