### 2017-08-12 786afe4
update ROOT/root_numpy comparison with latest ROOT and with both compiled with -O2
