### 2017-08-12 #307 @ndawe
Sure, thanks. I've recreated the benchmark plot after updating to the latest ROOT and with both root_numpy and ROOT compiled with -O2. The diff is here and you can see a slightly closer match:

https://github.com/scikit-hep/root_numpy/pull/346/files

Merging that now. Thanks @wlav for the thorough review!
