### 2017-08-12 #307 @danielskatz
Thanks again @wlav for a very nice review, and @ndawe for your updates.

@arfon - this is ready to go to you now.
