### 2017-08-14 #307 @ndawe
@arfon I've posted a new release on Zenodo: https://zenodo.org/record/842249

DOI: http://doi.org/10.5281/zenodo.842249

The codemeta.json is also updated
