### 2017-08-14 #307 @ndawe
@arfon the figure referencing seems to be broken on the web version of the paper ("... as shown in Figure ." and no "Figure #" with the caption). Can this be fixed?
