### 2017-08-15 #307 @arfon
@ndawe - apparently this is a known issue with Pandoc --> HTML conversion (https://stackoverflow.com/questions/9434536/how-do-i-make-a-reference-to-a-figure-in-markdown-using-pandoc).

If you can modify the language of the text to not directly cite the figure I can recompile the documents and update them.
