### 2017-08-16 #307 @ndawe
Thanks @arfon I've updated the text to remove the reference
